First draft of coping with py-fido2>=0.9.
Much of the API changed, as did -- unfortunately -- the serialization
of (Collected)ClientData objects in JSON.

Fortunately, most of the changes are cosmetic.

The serialization of (Collected)ClientData objects was not requested
in any canonical form by py-fido2's calls to json.dumps, so we need
some care to match the old behaviour in order to ensure newly created
signatures can be verified by existing verifiers, and the new
verifier can verify existing signatures.  (Perhaps nobody uses
fidosig, but I don't want to make that assumption!)  To be tidied up
a bit in a subsequent revision.